Looking for persistency partitions on luks devices only if we boot with persistent=cryptsetup.
In order to determine if a encrypted partition contains a persistent
partition for live-initramfs, we do have to open it and look into it.
Open means, that the user has to provide the passphase (or key).

A user booting a *default* debian-live system on a machine that has
cryptesetup encrypted partitions should not be bothered by *default*
to enter all his passphrases during the boot process. Also,
encrypted persistency is only usefull when the user knows about it
(= the passphrase to access it). Therefore, it's better to by default
ignore persistency on luks devices.