Erroring out in case a "wrong" filesystem was chosen.
If live-boot finds a "wrong" filesystem that looks OK, then the error message
can be pretty confusing. Panic with an explicit message in this case.

The "is this rootfs valid" test is implemented by checking if the
/usr/share/doc/live-boot/ directory exists: as per Debian Policy 2.3 (Copyright
considerations), this directory must exist on any system where the live-boot
binary package is installed.

Adapted and reformatted from a patch by Christian Hofstaedtler <ch@grml.org>.