Adding persistent-storage=TYPE...
The comma-separated list TYPE... specifies which types of persistent
storage to use. Current possible TYPEs are {file, filesystem}. For
instance, persistent-storage=filesystem is equivalent to the old
persistent=nofiles, but persistent-storage=file allows to look for
only files, which wasn't possible before. The default is to look for
both, i.e. persistent-storage=filesystem,file.