Remove cow dir contents that are in the way for custom mounts.
This is mostly relevant if you activate persistence post boot; in that
case you (or live-config) may have modified or added some file, which
makes it appear in the cow dir. If you later activate a persistent
volume which contains that file, the old file will "block" the one
from the persistent volume, which likely isn't what you expect.