Blacklist devices used for /live/rootfs/* insted of /live/medium.
The path /live/medium is used for probing different devices when searching
for the live rootfs. There is no reason why that device should be banned
from being used for persistence storage. Instead the devices that are
actually used as the backing storage for the live rootfs should be banned.