Do not "exit" if boot option nonetworking is set but "return" instead
Nowadays we're running under one main script which sources all
the further scripts. If we 'exit' in a script then we'd just exit
the main init sequence, resulting in a kernel trace because the
real™ init hasn't been found/executed.

Thanks to Florian Ermisch <floh.edo@googlemail.com> for the initial bug report