Update 9990-misc helper to support latest udev version
udevadm trigger-test was a debug interface and was never supposed to be used public. By using udevadm trigger-test in the is_nice_function proper DEVICES are no longer recognized. This patch addresses the issue by using udevadm info and checking if the DEVTYPE of the device is disk. This should work with the majority of DEVICES.

Please be aware that you probably shouldn't merge this patch directly as you should create a debian/patch/ file for it. This pull request is only for documentation purposes to document an workaround.