Update 9990-misc-helpers.sh to support latest udev version [Closes: issue1278]
When booting the ISO in hard-disk mode udev's
'udevadm test-builtin' emits this error message:

| calling: test-builtin
| error reading /lib/udev/hwdb.bin: No such file or directory
| load module index
| unload module index

and fails to boot because using udevadm test-builtin in the
is_nice_device function no longer recognizes proper DEVICES.

Including the /lib/udev/hwdb.bin file in the initramfs doesn't
help either as udevadm test-builtin was a debug interface and
never supposed to be used public.

This patch addresses the issue by using udevadm info and checking
if the DEVTYPE of the device is 'disk'. This should work with the
majority of DEVICES.

Closes: https://github.com/grml/live-boot-grml/pull/3

Thanks: Ulrich Dangel <mru@spamt.net> for the initial patch
Thanks: Michael Biebl <biebl@debian.org> for helping in resolve this issue