Supporting resolvconf in the squashfs.
When resolvconf is used in the squashfs, /etc/resolv.conf is a symlink
to the generated version. Depending on the size of the squashfs and
other factors, sometimes we try to write to /etc/resolv.conf while it
still points to nirvana, as resolvconf did not generate it yet.

Instead of being racy and writing to a file which will be regenerated
anyways, let's detect resolvconf and write to its base file instead.

Initial detection idea by Mika Prokop <mika@grml.org>