Some fixups to ensure we work with busybox/klibc-utils
Now that the PATH is sane again, no need to hardcode the path of many
tools.

Also it means that we use busybox/klibc-utils's mount and we should thus
aim to use a syntax that work with both of them. We thus use "mount -o
move" instead of "mount --move", and "mount -o bind" instead of "mount
--bind".