Use klibc's mount again for fuse mounts
When building a stretch live image which includes httpfs/buster for the
created live-image's initramfs to support live-boot's httpfs switch, the
boot process fails in a way similar to what has been reported in #823856.

Special handling for ${FUSE_MOUNT}s (httpfs, curlftps) was added to use
util-linux's mount instead of the klibc's in such cases. I tested the
use of a FUSE-based rootfs in conjunction with klibc's mount, and it
seems, nowadays the both of them work together.

So, the conditional incorporation and replacement of the mount command
is both no longer necessary, and has become harmful.

Closes: #868559