Add backward compatibility rbind mount /lib/live/mount -> /run/live
The paths used in the current released versions of live-boot are a
form of public API, and existing applications and scripts might rely
on them. Do a recursive bind mount of the new path on the previous one
so that they do not break on upgrade (see #886328).

This backward-compatible mount point will be deprecated and removed
before the Bullseye (Debian 11) release. Users are recommended to start
migrating to the new /run/live path as soon as possible.

Gbp-Dch: Full