Fix resolving FQDN given by DHCP
The DHCP can specify a host and domain name for the system. live-boot
fills `/etc/hostname` and `/etc/resolv.conf` with the data from the DHCP
server, but resolving the fully qualified domain name (FQDN) fails, when
the host is not specified in DNS:

```
(initramfs) hostname -f
hostname: example-host: Host name lookup failure
```

To make the FQDN resolvable without DNS, `libnss_files.so` needs to be
included in the initramfs, `/etc/nsswitch.conf` need to prefer `files`
over `dns`, and `/etc/hosts` needs to be filled with the data from the
DHCP server.