9990-misc-helpers.sh: support lvm/mdadm in fromiso feature.
Previously, it wasn't possible to use fromiso with a Linux software RAID
or LVM device.

The code handling lvm/mdadm devices has been moved around and the
fromiso feature split up, so that lvm/mdadm devices are now started as
necessary even when fromiso has been passed.

The usual caveats apply: if you start from an lvm/mdadm device *without*
using the toram feature, the original volume will be kept mounted, so
it's not safe to mount the volume containing the grml image after
booting it.

With the toram feature, this is not an issue, since all the data is read
into a ramdisk and the original volume unmounted.