9990-main.sh: activate LVM volumes if needed.
An update to the Debian lvm2 package removed a script in the initrd that
would activate any LVs that are required to access the root volume. Due
to this change, grml failed to find its live media when it was stored on
an LVM volume.

Add an explicit activation of any LVs mentioned in the `live-media` or
`boofrom` kernel arguments, so that the live media can be found again.

Fixes https://github.com/grml/grml/issues/200