live-netdev: unify coding style and drop noop $NETDEV assignment
Upstream uses tabs instead of whitespace and if\nthen etc coding style,
while I personally don't like this, let's keep coding style consistent,
to make synching and contributions towards upstream easier.

While at it, the NETDEV="$NETDEV" is a noop, as the variable is assigned
to itself, so drop it.